requirements.txt
torch
torchvision
pillow
wandb
datasets
tqdm
diffusers
transformers
torch-ema
tf-yarn
fsspec
